requirements.txt
django==4.2.7
django-crispy-forms==2.0
django-allauth==0.58.2
pillow==10.1.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.24
django-bootstrap4==23.2
django-notifications-hq==1.8.2
channels==4.0.0
whitenoise==6.6.0
django-environ==0.11.2 